fix: I added alerts on all forms with sweetalert2

+++ client/src/Components/ProductList/ProductsList.jsx
@@ -2,6 +2,7 @@ import React, {useEffect, useState} from 'react';
 import axios from 'axios';
 import EditProduct from '../EditProduct/EditProduct';
 import {Link} from 'react-router-dom'
+import Swal from 'sweetalert2'
 const {REACT_APP_BACKEND_URL} = process.env;
 
 
@@ -21,12 +22,31 @@ const ProductsList = (props) => {
     }
 
       const handleDelete = id => {
-        axios.delete(`${REACT_APP_BACKEND_URL}/products/${id}`)
-          .then(res =>{
-                alert('se elimino')
+        Swal.fire({
+          title: 'Esta seguro de eliminar el producto?',
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Si, eliminar!'
+        }).then((result) => {
+          if (result.isConfirmed) {
+            axios.delete(`${REACT_APP_BACKEND_URL}/products/${id}`)
+            .then(res =>{
                 setProducts(products.filter(p => p.id !== id))
-          })
+            })
+            Swal.fire({
+              title:'Eliminado!',
+              icon:'success'
+            })
+          }
+        })
+
+
+        
       }
+
+    
 
   
     return (
